requirements.txt
anthropic==0.68.1
click==8.1.7
python-dotenv==1.0.1
colorama==0.4.6
pydantic==2.8.2
requests==2.32.3